refactor: use int16_t for position warp target coordinates
Replace 'int' with 'int16_t' for target_x and target_y member variables
in CheatSystem to ensure consistency with the codebase's use of fixed-width
integer types. This makes the intended range and signedness explicit while
maintaining proper validation against std::atoi() return values.

+++ include/cheats.h
@@ -71,8 +71,8 @@ private:
     bool awaiting_x_input;
     bool awaiting_y_input;
     std::string position_input_buffer;
-    int target_x;
-    int target_y;
+    int16_t target_x;
+    int16_t target_y;
 };
 
 // Global cheat system instance
